Wrote algo for random dungeons
This algorithm can create a two-dimensional maze (dungeon, labyrinth) with rooms that have a minimum threshold of dimensions, and it can be customized in different ways. Still, It lacks of an implementation of corridors for making different rooms communicate. It is inspired by the Space Partitioning algorithm, that uses Space Partitioning Trees.